+++ pentagon/meta.py
@@ -1,3 +1,3 @@
 
-__version__ = "2.7.1"
+__version__ = "2.7.2"
 __author__ = 'ReactiveOps, Inc.'