+++ CHANGELOG.md
@@ -4,10 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## 2.7.2
 ### Added
 - Instructions on how to setup the development environment.
 - revised cli help text
+- migrations for kops settings that were missed in the last migration
+- made `anonymousAuth: false` default for Kops clusters. This currently conflicts with metricserver version > 3.0.0
+
+## 2.7.1
+### Fixed
+- migration
 
 
 ## 2.7.0 - 2019-1-3